Adapt truth analyzer for ttW
